requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
alembic==1.1.0            # via flask-migrate
aniso8601==8.0.0          # via flask-restful
attrs==19.1.0             # via packaging
authlib[client]==0.12.1
beautifulsoup4==4.8.0
blinker==1.4              # via flask-debugtoolbar, sentry-sdk
boto3==1.9.228
botocore==1.12.228        # via boto3, s3transfer
canvasapi==0.14.0
cas-client==1.0.0
certifi==2019.9.11        # via requests, sentry-sdk
cffi==1.13.1              # via cryptography
chardet==3.0.4            # via requests
click==7.0                # via flask, rq
croniter==0.3.30          # via rq-scheduler
cryptography==2.8         # via authlib
docutils==0.15.2          # via botocore
email-reply-parser==0.5.9
faker==2.0.1
filelock==3.0.12          # via tox
flask-admin==1.5.3
flask-caching==1.7.2
flask-debugtoolbar==0.10.1
flask-jwt-extended==3.23.0
flask-login==0.4.1
flask-marshmallow==0.10.1
flask-migrate==2.5.2
flask-restful==0.3.7
flask-rq2==18.3
flask-sqlalchemy==2.4.0   # via flask-migrate
flask-sslify==0.1.5
flask-wtf==0.14.2
flask==1.1.1              # via flask-admin, flask-caching, flask-debugtoolbar, flask-jwt-extended, flask-login, flask-marshmallow, flask-migrate, flask-restful, flask-rq2, flask-sqlalchemy, flask-sslify, flask-wtf, sentry-sdk
gunicorn==19.9.0
idna==2.8                 # via requests
importlib-metadata==0.22  # via pluggy, tox
itsdangerous==1.1.0       # via flask, flask-debugtoolbar
jinja2==2.10.1            # via flask
jmespath==0.9.4           # via boto3, botocore
mako==1.1.0               # via alembic
markupsafe==1.1.1         # via jinja2, mako
marshmallow-sqlalchemy==0.19.0
marshmallow==3.1.1        # via flask-marshmallow, marshmallow-sqlalchemy
more-itertools==7.2.0     # via zipp
packaging==19.1           # via tox
pluggy==0.13.0            # via tox
psycopg2-binary==2.8.3
py==1.8.0                 # via tox
pycparser==2.19           # via cffi
pycryptodome==3.9.0       # via cas-client
pyjwt==1.7.1              # via flask-jwt-extended
pyparsing==2.4.2          # via packaging
python-dateutil==2.8.0    # via alembic, botocore, croniter, faker
python-editor==1.0.4      # via alembic
python-http-client==3.2.1  # via sendgrid
pytz==2019.2              # via canvasapi, flask-restful
pyyaml==5.1.2
redis==3.3.8              # via flask-rq2, rq
requests==2.22.0          # via authlib, canvasapi, cas-client
rq-scheduler==0.9         # via flask-rq2
rq==1.1.0                 # via flask-rq2, rq-scheduler
s3transfer==0.2.1         # via boto3
sendgrid==6.1.0
sentry-sdk[flask]==0.12.2
six==1.12.0               # via canvasapi, cas-client, cryptography, faker, flask-jwt-extended, flask-marshmallow, flask-restful, packaging, python-dateutil, tox
soupsieve==1.9.3          # via beautifulsoup4
sqlalchemy==1.3.8         # via alembic, flask-sqlalchemy, marshmallow-sqlalchemy
text-unidecode==1.2       # via faker
toml==0.10.0              # via tox
tox==3.14.0               # via cas-client
urllib3==1.25.3           # via botocore, requests, sentry-sdk
virtualenv==16.7.5        # via tox
werkzeug==0.15.6          # via flask, flask-debugtoolbar, flask-jwt-extended
wtforms==2.2.1            # via flask-admin, flask-wtf
zipp==0.6.0               # via importlib-metadata
